+++ app/models.py
@@ -170,12 +170,19 @@ class Role(db.Model):
 
 
 class Punch(db.Model):
-    """Table: punches"""
+    '''Table: punches'''
     __tablename__ = 'punches'
     user_id = db.Column(db.Integer, db.ForeignKey('users.id'), primary_key=True)
     video_id = db.Column(db.Integer, db.ForeignKey('videos.id'), primary_key=True)
     play_time = db.Column(db.Interval, default=timedelta())
+    synchronized = db.Column(db.Boolean, default=False)
     timestamp = db.Column(db.DateTime, default=datetime.utcnow)
+
+    def set_synchronized(self):
+        '''Punch.set_synchronized(self)'''
+        self.synchronized = True
+        self.timestamp = datetime.utcnow()
+        db.session.add(self)
 
     @property
     def complete(self):
@@ -212,6 +219,7 @@ class Punch(db.Model):
             str(self.user_id),
             str(self.video_id),
             str(self.play_time.total_seconds()),
+            str(int(self.synchronized)),
             self.timestamp.strftime(current_app.config['DATETIME_FORMAT']),
         ]
         return csv_entry
@@ -231,8 +239,9 @@ class Punch(db.Model):
                             user_id=int(entry[0]),
                             video_id=int(entry[1]),
                             play_time=timedelta(seconds=float(entry[2])),
+                            synchronized=bool(int(entry[3])),
                             timestamp=datetime.strptime(
-                                entry[3],
+                                entry[4],
                                 current_app.config['DATETIME_FORMAT']
                             )
                         )
@@ -256,6 +265,7 @@ class Punch(db.Model):
                 'user_id',
                 'video_id',
                 'play_time',
+                'synchronized',
                 'timestamp',
             ])
             for entry in Punch.query.all():
@@ -430,26 +440,31 @@ class User(UserMixin, db.Model):
         }
         return json_entry
 
-    def punch(self, video, play_time=None):
-        '''User.punch(self, video, play_time=None)'''
+    def punch(self, video, play_time, synchronized=False):
+        '''User.punch(self, video, play_time, synchronized=False)'''
         punch = self.punches.filter_by(video_id=video.id).first()
         if punch is not None:
+            punch.synchronized = synchronized
             punch.timestamp = datetime.utcnow()
         else:
             punch = Punch(
                 user_id=self.id,
-                video_id=video.id
+                video_id=video.id,
+                synchronized=synchronized
             )
-        if play_time is not None:
-            if isinstance(play_time, timedelta):
-                punch.play_time = play_time
-            else:
-                punch.play_time = timedelta(seconds=play_time)
+        if isinstance(play_time, timedelta):
+            punch.play_time = play_time
+        else:
+            punch.play_time = timedelta(seconds=play_time)
         db.session.add(punch)
 
     def punched(self, video):
         '''User.punched(self, video)'''
-        return self.punches.filter_by(video_id=video.id).first() is not None
+        return self.get_punch(video=video) is not None
+
+    def get_punch(self, video):
+        '''User.get_punch(self, video)'''
+        return self.punches.filter_by(video_id=video.id).first()
 
     @property
     def latest_punch(self):
@@ -480,26 +495,39 @@ class User(UserMixin, db.Model):
             .order_by(Video.id.desc())\
             .first()
 
-    def lesson_progress(self, lesson):
-        '''User.lesson_progress(self, lesson)'''
+    def lesson_play_time(self, lesson):
+        '''User.lesson_play_time(self, lesson)'''
         return reduce(operator.add, [punch.play_time_trim for punch in Punch.query\
             .join(Video, Video.id == Punch.video_id)\
             .filter(Video.lesson_id == lesson.id)\
             .filter(Punch.user_id == self.id)\
-            .all()], timedelta()) / lesson.duration
+            .all()], timedelta())
+
+    def lesson_progress(self, lesson):
+        '''User.lesson_progress(self, lesson)'''
+        return self.lesson_play_time(lesson=lesson) / lesson.duration
 
     def lesson_progress_percentage(self, lesson):
         '''User.lesson_progress_percentage(self, lesson)'''
         return '{:.0%}'.format(self.lesson_progress(lesson=lesson))
 
-    def lesson_punch(self, lesson):
-        '''User.lesson_punch(self, lesson)'''
-        return Punch.query\
-            .join(Video, Video.id == Punch.video_id)\
-            .filter(Video.lesson_id == lesson.id)\
-            .filter(Punch.user_id == self.id)\
-            .order_by(Punch.timestamp.desc())\
-            .first()
+    def complete_lesson(self, lesson, progress_threshold=None):
+        '''User.complete_lesson(self, lesson, progress_threshold=None)'''
+        if progress_threshold is not None:
+            return self.lesson_progress(lesson=lesson) >= progress_threshold
+        return self.lesson_progress(lesson=lesson) >= lesson.progress_threshold
+
+    def can_study(self, lesson):
+        '''User.can_study(self, lesson)'''
+        if lesson.type.name in ['VB', 'Y-GRE', 'Y-GRE AW']:
+            return self.plays(role_name='协调员') or \
+                reduce(operator.and_, [self.complete_lesson(lesson=lesson) \
+                    for lesson in Lesson.query\
+                    .join(LessonType, LessonType.id == Lesson.type_id)\
+                    .filter(Lesson.type_id == lesson.type_id)\
+                    .filter(Lesson.id < lesson.id)\
+                    .all()], True)
+        return True
 
     def video_play_time(self, video):
         '''User.video_play_time(self, video)'''
@@ -519,27 +547,11 @@ class User(UserMixin, db.Model):
         '''User.video_progress_percentage(self, video)'''
         return '{:.0%}'.format(self.video_progress(video=video))
 
-    def complete_video(self, video):
-        '''User.complete_video(self, video)'''
-        punch = self.punches.filter_by(video_id=video.id).first()
-        return punch is not None and punch.complete
-
-    def complete_lesson(self, lesson):
-        '''User.complete_lesson(self, lesson)'''
-        return reduce(operator.and_, [self.complete_video(video=video) \
-            for video in lesson.videos], True)
-
-    def can_study(self, lesson):
-        '''User.can_study(self, lesson)'''
-        if lesson.type.name in ['VB', 'Y-GRE', 'Y-GRE AW']:
-            return self.plays(role_name='协调员') or \
-                reduce(operator.and_, [self.complete_video(video=video) for video in Video.query\
-                    .join(Lesson, Lesson.id == Video.lesson_id)\
-                    .join(LessonType, LessonType.id == Lesson.type_id)\
-                    .filter(Lesson.type_id == lesson.type_id)\
-                    .filter(Lesson.id < lesson.id)\
-                    .all()], True)
-        return True
+    def complete_video(self, video, progress_threshold=None):
+        '''User.complete_video(self, video, progress_threshold=None)'''
+        if progress_threshold is not None:
+            return self.video_progress(video=video) >= progress_threshold
+        return self.video_progress(video=video) >= video.lesson.progress_threshold
 
     def can_play(self, video):
         '''User.can_play(self, video)'''
@@ -1040,6 +1052,7 @@ class Lesson(db.Model):
     name = db.Column(db.Unicode(64), unique=True, index=True)
     abbr = db.Column(db.Unicode(64))
     type_id = db.Column(db.Integer, db.ForeignKey('lesson_types.id'))
+    progress_threshold = db.Column(db.Float, default=0.0)
     videos = db.relationship('Video', backref='lesson', lazy='dynamic')
 
     @property
@@ -1063,7 +1076,8 @@ class Lesson(db.Model):
                 lesson = Lesson(
                     name='{} {}'.format(entry['lesson_type_name'], entry['abbr']),
                     abbr=entry['abbr'],
-                    type_id=LessonType.query.filter_by(name=entry['lesson_type_name']).first().id
+                    type_id=LessonType.query.filter_by(name=entry['lesson_type_name']).first().id,
+                    progress_threshold=entry['progress_threshold']
                 )
                 db.session.add(lesson)
                 if verbose: