+++ RELEASE-NOTES.md
@@ -1,6 +1,6 @@
 ## RELEASE NOTES
 
-### Version 1.0.1 - January 21, 2019
+### Version 1.0.1 - January 22, 2019
 
 **NEW**
 - **Model** - Add `demo` attribute to video data model #1
@@ -18,6 +18,8 @@
 - **Feature** - Adopt Y-System user authentication API #32
 - **Feature** - Authenticate user via Y-System #37
 - **Feature** - Email production server errors to system operator #39
+- **Model** - Add `synchronized` attribute to punch data model #40
+- **Model** - Add `progress_threshold` attribute to lesson data model #41
 
 **IMPROVED**
 - **Model** - Change `demo` attribute of video data model to `restricted` #7